refactor: SharedQueryClient
